+++ packages/web-api-scan-runner/src/report-generator/axe-result-to-html-converter.ts
@@ -2,22 +2,11 @@
 // Licensed under the MIT License.
 
 import { AxeResults } from 'axe-core';
-import { convertAxeToSarif, SarifLog } from 'axe-sarif-converter';
 import { inject, injectable } from 'inversify';
 import { ReporterFactory } from 'markreay-accessibility-insights-report';
 import { ReportFormat } from 'storage-documents';
 import { iocTypeNames } from '../ioc-types';
-
-export type ReportGenerationParams = {
-    pageTitle: string;
-};
-
-@injectable()
-export abstract class AxeResultConverter {
-    public readonly reportType: ReportFormat;
-
-    public abstract convert(results: AxeResults, params: ReportGenerationParams): string;
-}
+import { AxeResultConverter, ReportGenerationParams } from './axe-result-converter';
 
 @injectable()
 export class AxeResultToHtmlConverter extends AxeResultConverter {
@@ -40,20 +29,5 @@ export class AxeResultToHtmlConverter extends AxeResultConverter {
         };
 
         return reporter.fromAxeResult(htmlReportParams).asHTML();
-    }
-}
-
-@injectable()
-export class AxeResultToSarifConverter extends AxeResultConverter {
-    public readonly reportType: ReportFormat = 'sarif';
-
-    constructor(@inject(iocTypeNames.ConvertAxeToSarifFunc) private readonly convertAxeToSarifFunc: (axeResults: AxeResults) => SarifLog) {
-        super();
-    }
-
-    public convert(results: AxeResults, params: ReportGenerationParams): string {
-        const sarifResults = this.convertAxeToSarifFunc(results);
-
-        return JSON.stringify(sarifResults);
     }
 }
