+++ packages/web-api-scan-runner/src/runner/runner.ts
@@ -1,6 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { AxeResults } from 'axe-core';
 import { GuidGenerator } from 'common';
 import { inject, injectable } from 'inversify';
 import { isNil } from 'lodash';
@@ -25,7 +24,6 @@ import { WebDriverTask } from '../tasks/web-driver-task';
 
 @injectable()
 export class Runner {
-
     constructor(
         @inject(GuidGenerator) private readonly guidGenerator: GuidGenerator,
         @inject(ScanMetadataConfig) private readonly scanMetadataConfig: ScanMetadataConfig,
