+++ packages/web-api-scan-runner/src/setup-web-api-scan-runner-container.ts
@@ -9,8 +9,8 @@ import { reporterFactory } from 'markreay-accessibility-insights-report';
 import { registerScannerToContainer } from 'scanner';
 import { registerServiceLibraryToContainer } from 'service-library';
 import { iocTypeNames } from './ioc-types';
-import { AxeResultToHtmlConverter, AxeResultToSarifConverter } from './report-generator/axe-result-converters';
-import { ReportGenerator } from './report-generator/report-generator';
+import { AxeResultToHtmlConverter } from './report-generator/axe-result-to-html-converter';
+import { AxeResultToSarifConverter } from './report-generator/axe-result-to-sarif-converter';
 
 export function setupWebApiScanRequestSenderContainer(): inversify.Container {
     const container = new inversify.Container({ autoBindInjectable: true });
@@ -22,10 +22,12 @@ export function setupWebApiScanRequestSenderContainer(): inversify.Container {
 
     container.bind(iocTypeNames.ConvertAxeToSarifFunc).toConstantValue(convertAxeToSarif);
     container.bind(iocTypeNames.ReporterFactory).toConstantValue(reporterFactory);
-    container.bind(iocTypeNames.AxeResultConverters).toConstantValue([
-        container.get<AxeResultToSarifConverter>(AxeResultToSarifConverter),
-        container.get<AxeResultToHtmlConverter>(AxeResultToHtmlConverter),
-    ]);
+    container
+        .bind(iocTypeNames.AxeResultConverters)
+        .toConstantValue([
+            container.get<AxeResultToSarifConverter>(AxeResultToSarifConverter),
+            container.get<AxeResultToHtmlConverter>(AxeResultToHtmlConverter),
+        ]);
 
     return container;
 }
